Changing use of client API to pool in tests

+++ index.js
@@ -1,25 +1,35 @@
-const {Client} = require("pg");
+const {Pool} = require("pg");
 
-let client;
-
-    if (process.env.NODE_ENV === "test") {
-        client = new Client("postgres://localhost:5432/costume_shop_db_test");
-        client.on("error", (error) => {
+const getPool = () => {
+    let pool;
+    if (process.env.NODE_ENV === 'test') {
+       pool = new Pool({
+            host: 'localhost',
+            port: 5432,
+            database: 'costume_shop_db_test'
+        });
+        pool.on("error", (error) => {
             console.error(error.stack())
         })
     } else {
-        client = new Client("postgres://localhost:5432/costume_shop_db_dev");
-        client.on("error", (error) => {
+        pool = new Pool({
+            host: 'localhost',
+            port: 5432,
+            database: 'costume_shop_db_dev'
+        });
+        pool.on("error", (error) => {
             console.error(error.stack())
         })
     }  
+    return pool;
+}
 
 
-const createTables = async () => {
-    await client.query(`
+const createTables = async (pool) => {
+    await pool.query(`
         DROP TABLE IF EXISTS costumes;    
     `)
-    await client.query(` 
+    await pool.query(` 
         CREATE TABLE costumes(
             id SERIAL PRIMARY KEY,
             name VARCHAR(80) NOT NULL,
@@ -40,50 +50,46 @@ const createTables = async () => {
             price FLOAT NOT NULL 
         );
     `)
-    // Selecting table created by me, which should just be 'costumes'
-    const table = await client.query(`
-        SELECT * FROM pg_catalog.pg_tables WHERE tableowner='darayazdani';
-    `)
-    console.log(table.rows[0])
-    // Returning the name of the table I created
-    return table.rows[0].tablename;
 }
 
 const createCostume = async (
-    costumeName, 
-    category, 
-    gender, 
-    size, 
-    type, 
-    stockCount, 
-    price
-    ) => {
-        const {rows:[costume]} = await client.query(`
-            INSERT INTO costumes(
-                name,
-                category,
-                gender,
-                size,
-                type,
-                stock_count,
-                price
-            ) VALUES ($1, $2, $3, $4, $5, $6, $7)
-            RETURNING *;   
-        `, [costumeName, category, gender, size, type, stockCount, price]
-        )
+    pool,
+    {
+        name, 
+        category, 
+        gender, 
+        size, 
+        type, 
+        stock_count, 
+        price
+    }) => {
+    const {rows:[costume]} = await pool.query(`
+        INSERT INTO costumes(
+            name,
+            category,
+            gender,
+            size,
+            type,
+            stock_count,
+            price
+        ) VALUES ($1, $2, $3, $4, $5, $6, $7)
+        RETURNING *;   
+    `, [name, category, gender, size, type, stock_count, price]
+    )
     return costume;
 }
 
-const getAllCostumes = async () => {
-    const {rows: costumes} = await client.query(`
+const getAllCostumes = async (pool) => {
+    const {rows: costumes} = await pool.query(`
         SELECT * FROM costumes;
     `)
     return costumes;
 }
 
-const getCostumeById = async (id) => {
-    const {rows:[costume]} = await client.query(`
-        SELECT  
+const getCostumeById = async (pool, id) => {
+    const {rows:[costume]} = await pool.query(`
+        SELECT
+            id,  
             name,
             category,
             gender,
@@ -93,21 +99,26 @@ const getCostumeById = async (id) => {
             price 
         FROM costumes
         WHERE id = $1;
-    `, [id]) 
+    `, [id])
+    if (costume === undefined) {
+        throw new Error(`Could not retrieve data because id provided (${id}) does not exist in table.`)
+    } 
     return costume;
 }
 
 const updateCostume = async (
-        id, 
-        costumeName, 
+    pool,
+    id,
+    {
+        name, 
         category,
         gender,
         size,
         type,
-        stockCount,
+        stock_count,
         price
-) => {
-    const {rows: [costume]} = await client.query(`
+    }) => {
+    const {rows: [costume]} = await pool.query(`
         UPDATE costumes
         SET 
             name = $1,
@@ -120,32 +131,45 @@ const updateCostume = async (
         WHERE id = $8
         RETURNING *;
     `, [
-        costumeName, 
+        name, 
         category, 
         gender, 
         size, 
         type, 
-        stockCount, 
+        stock_count, 
         price, 
         id
     ])
+    if (costume === undefined) {
+        throw new Error(`Could not update row because id provided (${id}) does not exist in table.`)
+    } 
     return costume;
 }
 
-const deleteCostumeById = async (id) => {
-    const {rows: [costume]} = await client.query(`
+const deleteCostumeById = async (pool, id) => {
+    const {rows: costumes} = await pool.query(`
+        SELECT * FROM costumes 
+        WHERE id = $1;
+    `, [id])
+    
+    if (costumes.length === 0) {
+        throw new Error(`Could not delete row because id provided (${id}) does not exist in table.`)
+    } 
+
+    const {rows: costume} = await pool.query(`
         DELETE FROM costumes
         WHERE id = $1;
     `, [id])
+    
     return costume;
 }
 
 module.exports = {
-    client,
     createTables,
     createCostume,
     getAllCostumes,
     getCostumeById,
     updateCostume,
-    deleteCostumeById
+    deleteCostumeById,
+    getPool
 }
